emacs: tree: support fold/unfold thread or sub-thread
This patch allow the user to fold/unfold a thread or a sub-thread in
the current tree buffer by pressing "t" key.

By default a string is displayed at the beginning of the overlay to
indicate that this thread/sub-thread is folded.
Pressing again "t" on a folded thread/sub-thread will unfold it.

This feature works accross all the queries which are in Tree View.

Signed-off-by: Julien Masson <massonju.eseo@gmail.com>